fix a couple of harmless warnings during wxOSX build
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@61124 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/osx/core/glgrab.cpp
@@ -146,8 +146,8 @@ CGImageRef grabViaOpenGL(CGDirectDisplayID display, CGRect srcRect)
     glReadBuffer(GL_FRONT);
     
     
-    width = srcRect.size.width;
-    height = srcRect.size.height;
+    width = (GLint)srcRect.size.width;
+    height = (GLint)srcRect.size.height;
     
     
     bytewidth = width * 4; // Assume 4 bytes/pixel for now
