+++ src/unix/evtloopunix.cpp
@@ -200,7 +200,8 @@ bool wxConsoleEventLoop::Pending() const
 
 bool wxConsoleEventLoop::Dispatch()
 {
-    DispatchTimeout(wxFDIODispatcher::TIMEOUT_INFINITE);
+    DispatchTimeout(static_cast<unsigned long>(
+        wxFDIODispatcher::TIMEOUT_INFINITE));
 
     return true;
 }
